// src/shared/types.d.ts
interface FilterData {
	key: string;
	label: string;
	value: string;
	type: string;
}